services/audit-ledger-service/requirements.txt
kafka-python
web3